% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfr_scrape_functions.R
\name{get_league_passing_by_year}
\alias{get_league_passing_by_year}
\title{Get league averages for passing from PFR by season.}
\usage{
get_league_passing_by_year()
}
\value{
a tibble of data from 1932 to present with league average passingt metrics
}
\description{
Get league averages for passing from PFR by season.
}
\examples{
# get_league_passing_by_year()
}
