% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfr_scrape_functions.R
\name{get_pfr_ids}
\alias{get_pfr_ids}
\title{Scrape all index pages on pro-football-reference.com.}
\usage{
get_pfr_ids()
}
\value{
a tibble. See the dictionary for pfr_ids for a full listing.
}
\description{
The scrape is polite: it initiates a total of 26 HTTP GET requests over
its life. The majority of the time is spent running regular expression
code to extract player IDs from the index pages. It is slow. The function
will try and guess the number of cores on the user's system to spawn multiple
R sessions, but the max number of sessions it will spawn is capped at 20. The
letter "B" will always be the slowest to parse. Expect at least 3 minutes for
the scrape to complete.
}
\examples{
# pfr_ids <- get_pfr_ids()
}
