% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfr_scrape_functions.R
\name{load_ids}
\alias{load_ids}
\title{Loads PFR IDs via the csv file located at https://www.pro-football-reference.com/short/inc/players_search_list.csv}
\usage{
load_ids()
}
\value{
a tibble of players names, pfr ids, and active or not binary flag and year started
}
\description{
Loads PFR IDs via the csv file located at https://www.pro-football-reference.com/short/inc/players_search_list.csv
}
\examples{
# load_ids()
}
