% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfr_scrape_functions.R
\name{scrape_pfr_player_ids_by_letter}
\alias{scrape_pfr_player_ids_by_letter}
\title{Scrape names and identifiers of all players whose last name starts with a
given letter from pro-football-reference.com}
\usage{
scrape_pfr_player_ids_by_letter(letter)
}
\arguments{
\item{letter}{The first letter a player's surname. Must be capitalized.}
}
\value{
a tibble of player names, ids, and the first letter of the player's surname
}
\description{
Scrape names and identifiers of all players whose last name starts with a
given letter from pro-football-reference.com
}
