% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfr_scrape_functions.R
\name{scrape_player_id}
\alias{scrape_player_id}
\title{Scrape a player's ID}
\usage{
scrape_player_id(merge_name)
}
\arguments{
\item{merge_name}{A player name. Omit suffixes and periods for best results.}
}
\value{
a character vector of the player id
}
\description{
Politely scrape a player's ID from their pro-football-reference.com player
page URL. Learn how to install RSelenium here: https://www.youtube.com/watch?v=GnpJujF9dBw&t=10s
}
\examples{
# scrape_player_id("joe montana")
}
\seealso{
https://www.sports-reference.com/bot-traffic.html
}
